refactor(app): group dashboard components into _dashboard/ subfolder
src/app/ root had 17+ loose files mixing Next.js routing conventions
(layout, page, not-found, robots, sitemap, icons) with
dashboard-specific
client components and server actions.

Move 8 dashboard-only files to src/app/_dashboard/ (Next.js private
folder convention — underscore prefix excludes it from routing):
copy-url-button, create-item-form, delete-item-button, edit-item-form,
item-actions, item-edit-section, open-form-button,
regenerate-link-button.

Update 8 import paths in page.tsx accordingly.

Also remove src/modules/item/ and src/modules/ui/ — both contained
only `export {}` with no consumers anywhere in the codebase.
